Consolidate mock appointment creation
