+++ src/applications/check-in/api/local-mock-api/mocks/v2/shared/index.js
@@ -1,0 +1,5 @@
+const get = require('./get');
+
+module.exports = {
+  get: { ...get },
+};
