+++ src/applications/check-in/day-of/tests/e2e/happy-path/everything.happy.path.cypress.spec.js
@@ -7,7 +7,7 @@ import NextOfKin from '../../../../tests/e2e/pages/NextOfKin';
 import EmergencyContact from '../../../../tests/e2e/pages/EmergencyContact';
 import Appointments from '../pages/Appointments';
 import Confirmation from '../pages/Confirmation';
-import checkInData from '../../../../api/local-mock-api/mocks/v2/check-in-data';
+import sharedData from '../../../../api/local-mock-api/mocks/v2/shared';
 
 describe('Check In Experience', () => {
   describe('everything path', () => {
@@ -25,17 +25,17 @@ describe('Check In Experience', () => {
       initializeCheckInDataPost.withSuccess();
       initializeDemographicsPatch.withSuccess();
 
-      const rv1 = checkInData.get.createMultipleAppointments();
-      const earliest = checkInData.get.createAppointment();
+      const rv1 = sharedData.get.createMultipleAppointments();
+      const earliest = sharedData.get.createAppointment();
       earliest.startTime = '2021-08-19T03:00:00';
-      const midday = checkInData.get.createAppointment();
+      const midday = sharedData.get.createAppointment();
       midday.startTime = '2021-08-19T13:00:00';
-      const latest = checkInData.get.createAppointment();
+      const latest = sharedData.get.createAppointment();
       latest.startTime = '2027-08-19T18:00:00';
       rv1.payload.appointments = [latest, earliest, midday];
 
-      const rv2 = checkInData.get.createMultipleAppointments();
-      const newLatest = checkInData.get.createAppointment();
+      const rv2 = sharedData.get.createMultipleAppointments();
+      const newLatest = sharedData.get.createAppointment();
       newLatest.startTime = '2027-08-19T17:00:00';
       rv2.payload.appointments = [newLatest, earliest, midday];
       const responses = [rv1, rv2];
